refactor: Схемы валидации перенесены в контроллеры

+++ src/controllers/metricsController.js
@@ -1,6 +1,61 @@
+import Joi from 'joi';
 import * as MetricService from '../services/metricService.js';
+import { METRIC_TYPES, VALIDATE_ERROR_MESSAGES } from '../utils/constants.js';
+
+const getMetricByIdSchema = Joi.object({
+  params: Joi.object({
+    id: Joi.number().min(1).required()
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
+
+const createMetricSchema = Joi.object({
+  body: Joi.object({
+    metric: Joi.object({
+      type: Joi.string().required().valid(...METRIC_TYPES),
+      name: Joi.string().required().min(1).max(1024),
+      indicators: Joi.array().required().items(Joi.object({
+        text: Joi.string().required().min(1).max(512),
+      }))
+    }).required()
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
+
+const updateMetricSchema = Joi.object({
+  params: Joi.object({
+    id: Joi.number().min(1).required()
+  }),
+  body: Joi.object({
+    metric: Joi.object({
+      type: Joi.string().valid(...METRIC_TYPES),
+      name: Joi.string().min(1).max(1024)
+    }).required().min(1)
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
+
+const deleteMetricSchema = Joi.object({
+  params: Joi.object({
+    id: Joi.number().min(1).required()
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
+
+const addIndicatorToMetricSchema = Joi.object({
+  params: Joi.object({
+    id: Joi.number().min(1).required()
+  }),
+  body: Joi.object({
+    indicator: Joi.object({
+      text: Joi.string().required()
+    }).required()
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
 
 export const getMetricById = (req, res) => {
+  const { error } = getMetricByIdSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   const metricId = req.params.id;
   const metric = MetricService.getMetricById(metricId);
   if (!metric) {
@@ -11,6 +66,12 @@ export const getMetricById = (req, res) => {
 }
 
 export const createMetric = (req, res) => {
+  const { error } = createMetricSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   const createdMetric = MetricService.createMetric(req.body.metric);
   if (!createdMetric) {
     res.status(422).json({ message: 'Ошибка при создании метрики' });
@@ -25,6 +86,12 @@ export const getMetrics = (req, res) => {
 }
 
 export const updateMetric = (req, res) => {
+  const { error } = updateMetricSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   const updatedMetric = MetricService.updateMetric(req.params.id, req.body.metric);
   if (!updatedMetric) {
     res.status(422).json({ message: 'Ошибка при обновлении метрики '});
@@ -33,6 +100,12 @@ export const updateMetric = (req, res) => {
 }
 
 export const deleteMetric = (req, res) => {
+  const { error } = deleteMetricSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   const metricId = req.params.id;
   const isDeleted = MetricService.deleteMetric(metricId);
   
@@ -44,6 +117,12 @@ export const deleteMetric = (req, res) => {
 }
 
 export const addIndicatorToMetric = (req, res) => {
+  const { error } = addIndicatorToMetricSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   try {
     const createdIndicator = MetricService.addIndicatorToMetric(req.params.id, req.body.indicator);
     res.status(201).json(createdIndicator);
