+++ src/controllers/objectsGroupsController.js
@@ -1,6 +1,69 @@
+import Joi from 'joi';
 import * as ObjectsGroupService from '../services/objectsGroupService.js';
+import { VALIDATE_ERROR_MESSAGES } from '../utils/constants.js';
+
+const getObjectsGroupByIdSchema = Joi.object({
+  params: Joi.object({
+    id: Joi.number().min(1).required()
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
+
+const updateObjectsGroupSchema = Joi.object({
+  params: Joi.object({
+    id: Joi.number().min(1).required()
+  }),
+  body: Joi.object({
+    objectsGroup: Joi.object({
+      name: Joi.string().min(1),
+      socialDamagePotencialScore: Joi.number(),
+      materialDamagePotencialScore: Joi.number()
+    }).required().min(1)
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
+
+const addMetricToObjectsGroupSchema = Joi.object({
+  params: Joi.object({
+    id: Joi.number().min(1).required()
+  }),
+  body: Joi.object({
+    metric: Joi.object({
+      id: Joi.number().min(1).required(),
+      indicators: Joi.array().required().min(2).items(Joi.object({
+        id: Joi.number().min(1).required(),
+        value: Joi.number().required(),
+        text: Joi.string()
+      })).required()
+    }).required()
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
+
+const removeMetricFromObjectsGroupSchema = Joi.object({
+  params: Joi.object({
+    objectsGroupId: Joi.number().min(1).required(),
+    metricId: Joi.number().min(1).required(),
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
+
+const updateIndicatorsValuesSchema = Joi.object({
+  params: Joi.object({
+    id: Joi.number().min(1).required(),
+  }),
+  body: Joi.object({
+    indicators: Joi.array().required().items(Joi.object({
+      id: Joi.number().min(1).required(),
+      text: Joi.string(),
+      value: Joi.number().required(),
+    })).required()
+  })
+}).messages(VALIDATE_ERROR_MESSAGES).unknown(true);
 
 export const addMetricToObjectsGroup = (req, res) => {
+  const { error } = addMetricToObjectsGroupSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   ObjectsGroupService.addMetricToObjectsGroup(req.params.id, req.body.metric);
   res.sendStatus(201);
 }
@@ -11,6 +74,12 @@ export const getObjectsGroupsList = (req, res) => {
 }
 
 export const getObjectsGroupById = (req, res) => {
+  const { error } = getObjectsGroupByIdSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   const objectsGroupId = req.params.id;
   const objectsGroup = ObjectsGroupService.getObjectsGroupById(objectsGroupId);
 
@@ -23,6 +92,12 @@ export const getObjectsGroupById = (req, res) => {
 }
 
 export const updateObjectsGroup = (req, res) => {
+  const { error } = updateObjectsGroupSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   const updatedObjectsGroup = ObjectsGroupService.updateObjectsGroup(req.params.id, req.body.objectsGroup);
   if (!updatedObjectsGroup) {
     res.status(422).json({ message: 'Группа объектов не обновлена' });
@@ -32,6 +107,12 @@ export const updateObjectsGroup = (req, res) => {
 }
 
 export const updateIndicatorsValues = (req, res) => {
+  const { error } = updateIndicatorsValuesSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   const isUpdated = ObjectsGroupService.updateIndicatorsValues(req.params.id, req.body.indicators);
   if (isUpdated) {
     res.status(201).json(req.body.indicators);
@@ -41,6 +122,12 @@ export const updateIndicatorsValues = (req, res) => {
 }
 
 export const removeMetricFromObjectsGroup = (req, res) => {
+  const { error } = removeMetricFromObjectsGroupSchema.validate(req);
+  if (error !== undefined) {
+    res.status(400).json({ message: error.details[0].message });
+    return
+  }
+
   const isRemoved = ObjectsGroupService.removeMetric(req.params.objectsGroupId, req.params.metricId);
   if (isRemoved) {
     res.status(200).json({ id: req.params.metricId });
